Support different deployment scopes (group and subscription) (#9)
* Support different deployment scopes (resource group and subscription for now)

* Bug fixes and readme updates

+++ entrypoint.sh
@@ -14,11 +14,15 @@ config_path="$entrypoint_dir/config.json"
 envs_path="$entrypoint_dir/envs.json"
 secrets_path="$entrypoint_dir/secrets.json"
 
-# Extract provisioner configuration
 name_prefix=$(jq -r '.md_metadata.name_prefix' "$params_path")
-region=$(jq -r '.region // "eastus"' "$config_path")
-resource_group=$(jq -r --arg name_prefix "$name_prefix" '.resource_group // $name_prefix' "$config_path")
+
+# Extract provisioner configuration
+scope=$(jq -r '.scope // "group"' "$config_path")
+location=$(jq -r '.location // .region // "eastus"' "$config_path")
 complete=$(jq -r '.complete // true' "$config_path")
+
+# resource group scope settings
+resource_group=$(jq -r --arg name_prefix "$name_prefix" '.resource_group // $name_prefix' "$config_path")
 create_resource_group=$(jq -r '.create_resource_group // true' "$config_path")
 delete_resource_group=$(jq -r '.delete_resource_group // true' "$config_path")
 
@@ -85,11 +89,29 @@ fi
 az account set --subscription "$azure_subscription_id"
 echo "${GREEN}Authentication successful.${NC}"
 
+az_flags=""
+# Set command and flag settings based on scope
+case "$scope" in
+  group)
+    echo "Targeting resource group $resource_group"
+    az_flags+=" --resource-group $resource_group"
+    ;;
+  sub)
+    echo "Targeting subscription $azure_subscription_id"
+    az_flags+=" --location $location"
+    ;;
+  *)
+    echo -e "${RED}Error: Unsupported scope '$scope'. Expected 'group' or 'sub'.${NC}"
+    exit 1
+    ;;
+esac
+
 # Extract flags from provisioner config
-flags=""
 if [ "$complete" = "true" ]; then
-  flags="--mode Complete"
+  az_flags+=" --mode Complete"
 fi
+
+deployment_name="${name_prefix}-${MASSDRIVER_STEP_PATH}"
 
 # Handle deployment actions
 case "$MASSDRIVER_DEPLOYMENT_ACTION" in
@@ -97,29 +119,31 @@ case "$MASSDRIVER_DEPLOYMENT_ACTION" in
   plan)
     evaluate_checkov
     echo "Executing plan..."
-    az deployment group what-if $flags --name "$resource_group-$MASSDRIVER_STEP_PATH" --resource-group "$resource_group" --template-file template.bicep --parameters @params.json --parameters @connections.json
+    az deployment $scope what-if $az_flags --name $deployment_name --template-file template.bicep --parameters @params.json --parameters @connections.json
     ;;
 
   provision)
     evaluate_checkov
     echo "Provisioning resources..."
 
-    if [ "$create_resource_group" = "true" ]; then
-      echo "Creating resource group $resource_group in region $region..."
-      az group create --name "$resource_group" --location "$region"
-      echo "Resource group $resource_group created."
-    else
-      echo "Checking if resource group $resource_group exists..."
-      if az group exists --name "$resource_group" | grep -q "true"; then
-        echo "Resource group exists! Using existing resource group $resource_group"
+    if [ "$scope" = "group" ]; then
+      if [ "$create_resource_group" = "true" ]; then
+        echo "Creating resource group $resource_group in location $location..."
+        az group create --name "$resource_group" --location "$location"
+        echo "Resource group $resource_group created."
       else
-        echo -e "${RED}Error: Resource group $resource_group does not exist. If 'create_resource_group' is false, the resource group must already exist in Azure. To avoid this error, set 'create_resource_group' to 'true' in the provisioner configuration, or create the resource group $resource_group before provisioning.${NC}"
-        exit 1
+        echo "Checking if resource group $resource_group exists..."
+        if az group exists --name "$resource_group" | grep -q "true"; then
+          echo "Resource group exists! Using existing resource group $resource_group"
+        else
+          echo -e "${RED}Error: Resource group $resource_group does not exist. If 'create_resource_group' is false, the resource group must already exist in Azure. To avoid this error, set 'create_resource_group' to 'true' in the provisioner configuration, or create the resource group $resource_group before provisioning.${NC}"
+          exit 1
+        fi
       fi
     fi
 
     echo "Deploying bundle"
-    az deployment group create $flags --name "$resource_group-$MASSDRIVER_STEP_PATH" --resource-group "$resource_group" --template-file template.bicep --parameters @params.json --parameters @connections.json | tee outputs.json
+    az deployment $scope create $az_flags --name $deployment_name --template-file template.bicep --parameters @params.json --parameters @connections.json | tee outputs.json
 
     jq -s '{params:.[0],connections:.[1],envs:.[2],secrets:.[3],outputs:.[4].properties.outputs}' "$params_path" "$connections_path" "$envs_path" "$secrets_path" outputs.json > artifact_inputs.json
     for artifact_file in artifact_*.jq; do
@@ -133,14 +157,14 @@ case "$MASSDRIVER_DEPLOYMENT_ACTION" in
   decommission)
     echo "Decommissioning resources..."
     touch empty.bicep
-    az deployment group create $flags --name "$resource_group-$MASSDRIVER_STEP_PATH" --resource-group "$resource_group" --template-file empty.bicep
+    az deployment $scope create $az_flags --name $deployment_name --template-file empty.bicep
     rm empty.bicep
 
-    echo "Deleting deployment group $resource_group-$MASSDRIVER_STEP_PATH"
-    az deployment group delete --name "$resource_group-$MASSDRIVER_STEP_PATH" --resource-group "$resource_group"
+    echo "Deleting deployment group $deployment_name"
+    az deployment $scope delete --name "$deployment_name"
 
-    if [ "$delete_resource_group" = "true" ]; then
-      echo "Deleting resource group $resource_group in region $region"
+    if [ "$scope" = "group" ] && [ "$delete_resource_group" = "true" ]; then
+      echo "Deleting resource group $resource_group in location $location"
       az group delete --name "$resource_group" --yes
     fi
